Simultaneously support STT and TTS for Cocoapods
Needed to figure out a way to allow both STT and TTS in the same app without conflicts with their library dependencies (libogg.a and libopus.a)
